Move pytest file and notebook to same dir
